# feat: Cross-platform path handling enhancement with async optimization