… ENHANCEMENT_DOCUMENTATION.md
… VALIDATION_REPORT.md
… benchmarks/performance_comparison.py
… tests/core/test_async_module.py
… tests/core/utils/test_cross_platform.py