@PreistlyPython 2025-09-22
Closing this PR in favor of the simpler, more focused PR #1137.

**Reason for consolidation:**
- This PR adds 2233 lines of code for what should be a simple pathlib migration
- Introduces complexity with new async optimization features beyond the scope of issue #933
- @pUrGe12 appears to prefer simpler approaches based on review patterns
- PR #1137 provides a clean, minimal solution that directly addresses the core path compatibility issue

**The core requirement (issue #933) is:**
> "Refactor the code to make sure os path related logic is OS agnostic"

PR #1137 achieves this with minimal changes and enhanced security, following the KISS principle.

Will focus efforts on PR #1137 for faster review and merge. Thanks! 🚀